- Fixed latest commit for MSVC. - Went warning hunting. I think 6 are left in the whole thing. - Found MSVC's PI.
git-svn-id: file:///var/lib/svn/sakusen/cpp/trunk@338 6f43db7c-3df8-0310-95f7-eae406835aab

+++ libsakusen/oarchive.cpp
@@ -105,7 +105,7 @@ OArchive& OArchive::operator<<(const double& d)
 
 OArchive& OArchive::operator<<(const String& s)
 {
-  uint32 sLength = s.length();
+  uint32 sLength = (uint32)s.length();
   *this << sLength;
   ensureSpace(sLength);
   memcpy(buffer+length, s.c_str(), sLength);
